Fix for filters not persisting
